refactor: revamped folder structure of handler, shouldve done it in multiple commits
