+++ src/handler/auth/authenticated.rs
@@ -1,0 +1,5 @@
+mod revoke_token;
+mod revoke_all_tokens;
+
+pub use revoke_token::*;
+pub use revoke_all_tokens::*;
