+++ src/handler/chat/authenticated.rs
@@ -1,0 +1,7 @@
+mod get_chat;
+mod create_chat;
+mod add_users_to_chat;
+
+pub use get_chat::*;
+pub use create_chat::*;
+pub use add_users_to_chat::*;
