+++ src/handler/message/authenticated.rs
@@ -1,0 +1,7 @@
+mod create_message;
+mod update_message;
+mod get_messages;
+
+pub use create_message::*;
+pub use update_message::*;
+pub use get_messages::*;
