+++ src/handler/relation/authenticated.rs
@@ -1,0 +1,11 @@
+mod add_blocked;
+mod add_friend;
+mod confirm_friend;
+mod remove_blocked;
+mod remove_friend;
+
+pub use add_blocked::*;
+pub use add_friend::*;
+pub use confirm_friend::*;
+pub use remove_blocked::*;
+pub use remove_friend::*;
