+++ src/handler/server.rs
@@ -1,90 +1,1 @@
-use std::sync::Arc;
-use axum::{extract::{Path, State}, response::IntoResponse, Json};
-use serde::Deserialize;
-
-use crate::{dto::ServerGetResponse, model::{error, server, AppState}};
-use crate::middleware::auth::Ctx;
-use crate::dto::{ObjectToDTO, ServerCreateResponse};
-
-
-pub async fn get_server_auth(
-    State(state): State<Arc<AppState>>,
-    ctx: Ctx,
-    Path(server_id): Path<String>
-) -> impl IntoResponse
-{
-    let repo_server = &state.server;
-
-    let server = repo_server
-        .get_server_by_id(&server_id)
-        .await?;
-
-    let ctx_user_id = ctx.user_id_ref();
-    
-    if !server.is_user_part_of_server(ctx_user_id)
-    {
-        return Err(error::Server::ServerDoesNotContainThisUser);
-    }
-
-    Ok(Json(ServerGetResponse::obj_to_dto(server)))
-}
-
-#[derive(Deserialize)]
-pub struct CreateServerRequest
-{
-    name: String,
-}
-pub async fn create_server_auth(
-    State(state): State<Arc<AppState>>,
-    ctx: Ctx,
-    Json(payload): Json<CreateServerRequest>
-) -> impl IntoResponse
-{
-    let repo_server = &state.server;
-    let repo_user = &state.user;
-
-    let ctx_user_id = &ctx.user_id();
-
-    let owner = repo_user
-        .get_user_by_id(ctx_user_id)
-        .await?;
-
-    let server = server::Server::new(payload.name, owner)?;
-
-    match repo_server.create_server(server).await 
-    {
-        Ok(server) => Ok(Json(ServerCreateResponse::obj_to_dto(server))),
-        Err(e) => Err(e),
-    }
-}
-
-pub async fn join_server_auth(
-    State(state): State<Arc<AppState>>,
-    ctx: Ctx,
-    Path(server_id): Path<String>,
-) -> impl IntoResponse
-{
-    let repo_user = &state.user;
-    let repo_server = &state.server;
-
-    //add checks if joining is allowed
-    //like when rules get introduced (never)
-
-    let ctx_user_id = ctx.user_id_ref();
-
-    let user = repo_user
-        .get_user_by_id(ctx_user_id)
-        .await?;
-
-    let mut server = repo_server
-        .get_server_by_id(&server_id)
-        .await?;
-
-    server.add_user(user)?;
-
-    match repo_server.add_user_to_server(&server_id, ctx_user_id).await
-    {
-        Ok(()) => Ok(()),
-        Err(_) => Err(error::Server::FailedToAddUserToServer),
-    }
-}
+pub mod authenticated;
