+++ src/handler/user/admin.rs
@@ -1,0 +1,5 @@
+mod get_user;
+mod get_users;
+
+pub use get_user::*;
+pub use get_users::*;
