+++ src/handler/user/authenticated.rs
@@ -1,0 +1,3 @@
+mod get_ctx_user;
+
+pub use get_ctx_user::*;
